# documentation/swagger.yaml
...
swagger: '2.0'
info:
  title: 1ClickTrips API
  description: Business travel at its best
  version: "1.0.0"
# the domain of the service
host: api.1clicktrips.com
...
schemes:
  - http
  - https
# will be prefixed to all paths
basePath: /v1
produces:
  - application/json
paths:
...
  /auth/register:
    post:
      summary: User Signup
      description: |
        The user signup takes a username and a password as post arguments
        and returns an auth token with its expiration date.
      tags:
        - Auth
      parameters:
        - name: email
          in: body
          required: true
          schema:
            type: string
        - name: password
          in: body
          required: true
          schema:
            type: string
      responses:
        200:
          description: Authentication token for a user
          schema:
            $ref: '#/definitions/Authentication'
        401:
          description: Authentication failure
          schema: 
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /auth/local:
    post:
      summary: User Signin
      description: |
        The user signin takes a username and a password as post arguments
        and returns an auth token.
      tags:
        - Auth
      parameters:
        - name: email
          in: body
          required: true
          schema:
            type: string
        - name: password
          in: body
          required: true
          schema:
            type: string
      responses:
        200:
          description: Authentication token for a user
          schema:
            $ref: '#/definitions/Authentication'
        401:
          description: Authentication failure
          schema: 
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /auth/twitter:
    get: 
      summary: Twitter Signin
      description: |
        The Twitter authentication will be done with Twitter oauth2.
      tags:
        - Auth
      responses:
        200:
          description: Twitter signin successful
          schema:
            $ref: '#/definitions/Authentication'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  
#  /auth/connect/twitter:
#    get:
#      summary: Twitter Auth
#      description: |
#        Connect the Twitter account to the local account.
#      tags:
#        - Auth
#      responses: 
#        200:
#          description: Twitter connect successful
#          schema:
#            $ref: '#/definitions/Authentication'
#        default:
#          description: Unexpected error
#          schema:
#            $ref: '#/definitions/Error'
  /auth/unlink/twitter:
    get:
      summary: Unlink Twitter Accout
      description: |
        The Twitter account can be unlinked from the local account. 
      tags:
        - Auth
      responses:
        200:
          description: Twitter unlink successful
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /auth/logout:
    get:
      summary: Log out 
      description: |
        Log out the user and delete the cookie from the server.
      tags: 
        - Auth
        - User
      responses:
        200:
          description: Logout successful
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
            
  #/auth/google:
  #/auth/unlink/google:
    
  #/auth/facebook:
  #/auth/unlink/facebook:
            
  #########################################################################
  # User Account API
  #########################################################################
...
  /locales:
    get:
      summary: Language File
      description: | 
        The language file for a given language code.
      tags:
        - Language
        - Localization
      responses:
        200:
          description: A JSON array with the available languages
          schema:
            type: array
            items:
              $ref: '#/definitions/Language'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /locales/{languageCode}:
    get:
      summary: Language File
      description: | 
        The language file for a given language code.
      tags:
        - Language
        - Localization
      parameters:
        - name: languageCode
          in: path
          required: true
          type: string
          description: ISO 639-1 Language Codes
      responses:
        200:
          description: A JSON object with the translations
          schema:
            type: object
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
            
  #########################################################################
  # Currency API
  #########################################################################
  /currencies:
    get:
      summary: List of Currencies
      description: | 
        The list of available currencies.
      tags:
        - Language
        - Localization
      responses:
        200:
          description: A JSON array with the available currencies
          schema:
            type: array
            items:
              $ref: '#/definitions/Currency'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
            
#########################################################################
# API Data Definitions
#########################################################################
definitions:
  Authentication:
    required:
      - token
      - expires
    properties:
      token:
        type: string
        description: Authentication token
      expires:
        type: integer
        format: int32
        description: Expiration date of the auth token
  Language:
    properties:
      code: 
        type: string
        description: ISO 639-1 Language Code (e.g. en-US)
      name: 
        type: string
        description: The name of the language (e.g. English)
  Currency:
    properties:
      code:
        type: string
        description: EUR
      symbol: 
        type: string
        description: €
      thousandsSeparator: 
        type: string
      decimalSeparator: 
        type: integer
        format: int16
      symbolOnLeft: 
        type: boolean
      spaceBetweenAmountAndSymbol: 
        type: boolean
      roundingCoefficient: 
        type: integer
        format: int16
      decimalDigits: 
        type: integer
        format: int16
...
  Error:
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
      fields:
        type: string